+++ infra/tf/k8s_infra/nats.tf
@@ -2,9 +2,8 @@ locals {
 	service_nats = lookup(var.services, "nats", {
 		count = var.deploy_method_cluster ? 3 : 1
 		resources = {
-			cpu = 100
-			cpu_cores = 0
-			memory = 1000
+			cpu = 1000
+			memory = 1024
 		}
 	})
 }
@@ -38,6 +37,20 @@ resource "helm_release" "nats" {
 		podTemplate = {
 			merge = {
 				priorityClassName = kubernetes_priority_class.nats_priority.metadata.0.name
+			}
+		}
+		container = {
+			env = {
+				# See https://artifacthub.io/packages/helm/grafana/grafana#nats-container-resources
+				GOMEMLIMIT = "${floor(local.service_nats.resources.memory * 0.9)}MiB"
+			}
+			merge = {
+				resources = var.limit_resources ? {
+					limits = {
+						cpu = "${local.service_nats.resources.cpu}m"
+						memory = "${local.service_nats.resources.memory}Mi"
+					}
+				} : null
 			}
 		}
 		promExporter = {