+++ svc/pkg/captcha/ops/verify/src/lib.rs
@@ -74,9 +74,10 @@ async fn handle(
 			let origin_host = unwrap_ref!(ctx.origin_host, "no origin");
 
 			// Check for "rivet.game" host
-			let secret_key = if "rivet.game" == origin_host || origin_host.ends_with(".rivet.game")
-			{
-				Some(util::env::read_secret(&["turnstile", "rivet_game", "secret_key"]).await?)
+			let secret_key = if util::env::domain_cdn().map_or(false, |domain_cdn| {
+				domain_cdn == origin_host || origin_host.ends_with(&format!(".{domain_cdn}"))
+			}) {
+				Some(util::env::read_secret(&["turnstile", "cdn", "secret_key"]).await?)
 			}
 			// Check for host from captcha config
 			else {